backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
yt-dlp>=2023.10.13
moviepy>=1.0.3
ffmpeg-python>=0.2.0 